Move MeetupModal into it's own component file
